backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi>=23.1.0
bcrypt>=4.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0
requests==2.31.0
slowapi==0.1.9

# Resume template system
pdfplumber>=0.10.0
python-docx>=1.1.0
weasyprint>=60.0
Jinja2>=3.1.0
